2024-01-20 15:29 new commit

+++ finalproject/src/main/java/com/example/finalproject/finalproject/repository/RoadDamageRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.example.finalproject.finalproject.entity.RoadDamageEntity;
 
 @Repository
-public interface RoadDamageRepository extends JpaRepository<RoadDamageEntity, Long> {
+public interface RoadDamageRepository extends JpaRepository<RoadDamageEntity, Integer> {
     // 기본적인 CRUD 메서드는 JpaRepository에서 자동으로 제공됩니다.
 
 }
